+++ app/Http/Resources/LinkResource.php
@@ -13,11 +13,11 @@ class LinkResource extends JsonResource
             'title' => $this->title,
             'content' => $this->content,
             'url' => $this->url,
-            'is_health_check_enabled' => $this->is_health_check_enabled,
-            'http_status' => $this->getStatusText($this->http_status),
-            'http_status_color' => $this->getStatusColor($this->http_status),
-            'http_checked_at' => $this->http_checked_at,
+            'http_status' => $this->http_status,
+            'http_status_color' => $this->getHttpStatusColor($this->http_status),
+            'http_checked_at_formated' => $this->http_checked_at ? $this->http_checked_at->diffForHumans() : '',
             'permalink' => $this->permalink,
+            'is_watched' => $this->is_watched,
             'is_private' => $this->post->is_private,
             'is_pinned' => $this->post->is_pinned,
             'preview' => $this->preview,
@@ -38,33 +38,19 @@ class LinkResource extends JsonResource
         ];
     }
 
-    public function getStatusText($status)
+    protected function getHttpStatusColor($status): ?string
     {
-        if (is_null($status)) {
-            return null;
+        $range = (int)floor((int)$status / 100);
+
+        switch ($range) {
+            case 3:
+                return 'info';
+            case 4:
+                return 'warning';
+            case 5:
+                return 'danger';
+            default:
+                return null;
         }
-
-        if ($status == 200) {
-            return 'Healthy (200)';
-        }
-
-        if (400 <= $status and $status <= 499) {
-            return 'Dead (4xx)';
-        }
-
-        return 'Other (3xx, 5xx)';
-    }
-
-    public function getStatusColor($status)
-    {
-        if ($status == 200) {
-            return 'success';
-        }
-
-        if (400 <= $status and $status <= 499) {
-            return 'danger';
-        }
-
-        return 'warning';
     }
 }