+++ resources/views/manage/links.blade.php
@@ -1,5 +1,6 @@
 @extends('layouts.manage')
 
 @section('content')
-    @include('manage.links_dashboard')
+    <health-checks :enabled="{{ $enabled ? 'true' : 'false' }}"
+                   :stats="{{ json_encode($stats) }}"></health-checks>
 @endsection